make the reference time given on CL the zero-time
i.e. in the plot 0.0s (origin of time/x-axis) is actually the time given
on command line with option '-t'.
Before if an offset was given on command line, the 0-time in the plot
was reference time shifted by offset.
This helps easily read travel times if an origin time was given as
reference time on command line.