respect current filter settings when loading event and setting amplitude picks
without this, amplitude picks are set on the raw and unfiltered data and
thus usually not set very well. currently only the time of the min and
max of the peak to peak amplitude are stored in xml and when loading the
event the amplitude value of the data at these times is taken, so the
only real way to get amplitude picks right is when filter settings are
the same at the time of picking and when loading the event again and
when filter settings are respected during loading the event, which is
what this commit does.